DEXTORA/requirements.txt
# Web & API
fastapi[standard]==0.115.0
uvicorn==0.30.0
websockets==12.0

# AI & Reinforcement Learning
torch --index-url https://download.pytorch.org/whl/cpu
stable-baselines3==2.3.0
gymnasium==0.29.1
numpy>=1.24.0

# Database & Utilities
redis==5.0.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
sqlalchemy
psycopg[binary]
pgvector
google-cloud-storage